requirements.txt
tqdm==4.56.0
numpy==1.19.5
lxml==4.6.2
xmljson==0.2.1
requests==2.24.0
Pillow==8.0.1
matplotlib==3.3.2
scikit-image==0.17.2
pycocotools==2.0.2
opencv-python==4.4.0.44
pandas==1.1.4
scikit-learn==0.23.2
Deprecated==1.2.11
